# BookLibraryApi/bin/Release/netcoreapp3.1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "BookLibraryApi",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Genres/{GenreId}/Books": {
      "get": {
        "tags": [
          "Book"
        ],
        "operationId": "GetBooks",
        "parameters": [
          {
            "in": "path",
            "name": "GenreId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "SortBy",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Author",
            "type": "string"
          },
          {
            "in": "query",
            "name": "YearOfPublish",
            "type": "string"
          },
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "Fields",
            "type": "string"
          },
          {
            "in": "query",
            "name": "SearchQuery",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "head": {
        "tags": [
          "Book"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "GenreId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "SortBy",
            "type": "string"
          },
          {
            "in": "query",
            "name": "Author",
            "type": "string"
          },
          {
            "in": "query",
            "name": "YearOfPublish",
            "type": "string"
          },
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "Fields",
            "type": "string"
          },
          {
            "in": "query",
            "name": "SearchQuery",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Book"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "GenreId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BookCreation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "options": {
        "tags": [
          "Book"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "GenreId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BookCreation": {
      "required": [
        "bookTitle",
        "description"
      ],
      "type": "object",
      "properties": {
        "bookTitle": {
          "type": "string"
        },
        "bookCover": {
          "format": "uri",
          "type": "string"
        },
        "dateOfPublish": {
          "format": "date-time",
          "type": "string"
        },
        "numberOfBookPages": {
          "format": "int32",
          "type": "integer"
        },
        "publisher": {
          "type": "string"
        },
        "description": {
          "maxLength": 1000,
          "type": "string"
        },
        "authorId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
...
  }
}